requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dateparser==1.1.8
fastapi==0.103.1
fastjsonschema==2.18.0
fonttools==4.42.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
jieba==0.42.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
matplotlib==3.8.0
motor==3.3.1
multidict==6.0.4
NCHU-nlptoolkit==2.0.3
nltk==3.8.1
numpy==1.26.0
openai==1.44.1
packaging==23.1
Pillow==10.0.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tqdm==4.66.1
typing_extensions>=4.11,<5
tzlocal==5.0.1
udicOpenData==2.4
uri-template==1.3.0
urllib3==1.26.17
uvicorn==0.23.2
wordcloud==1.9.2
yarl==1.9.2
